Navigation.kt 按区域拆成七个 entry 文件；更新提醒与 Telegram 绑定各给 ViewModel；selfUid 改 StateFlow
1627 行的 Navigation.kt 里 700 行是一个 entryProvider 块，且有五处绕过
ViewModel 直摸仓库——包括从导航 lambda 发网络请求（WebView 的 Telegram
绑定轮询）。三件事一起还：

- entries 按区域拆进 TabEntries/SettingsEntries/AccountEntries/SpaceEntries/
  ToolsEntries/ThreadEntries/SessionEntries，共享捕获显式化为 StackEntryScope
  ——每栈一个实例，在 destinationProvider 里构建，保住「闭包必须绑定本栈」
  的既有约束（scope 的 KDoc 记了为什么；会变的值以函数过境，避免建 entry
  时冻结）。
- UpdateReminderViewModel 接管启动提醒对话框和我的页的小红点；
  TelegramBindingViewModel 接管绑定轮询的请求与「失败=还没绑上」策略。
  导航层不再有任何仓库直调。
- ProfileRepository.selfUid 从 @Volatile var 改成 StateFlow：点击时读
  .value 与原语义一致，组合将来要响应它也有得收。四处 isSelf 判断收敛
  到 StackEntryScope.openSpace 一个 lambda。
